Feature: general UI improvements (#35)
Confirmation dialogs, error and success messages

+++ src/modules/edc-demo/components/contract-definition-viewer/contract-definition-viewer.component.ts
@@ -2,8 +2,12 @@ import {Component, OnInit} from '@angular/core';
 import {BehaviorSubject, Observable, of} from 'rxjs';
 import {first, map, switchMap} from 'rxjs/operators';
 import {MatDialog} from '@angular/material/dialog';
-import {ContractDefinitionEditorDialog} from '../contract-definition-editor-dialog/contract-definition-editor-dialog.component';
+import {
+  ContractDefinitionEditorDialog
+} from '../contract-definition-editor-dialog/contract-definition-editor-dialog.component';
 import {ContractDefinitionDto, ContractDefinitionService} from "../../../edc-dmgmt-client";
+import {ConfirmationDialogComponent, ConfirmDialogModel} from "../confirmation-dialog/confirmation-dialog.component";
+import {NotificationService} from "../../services/notification.service";
 
 
 @Component({
@@ -17,7 +21,9 @@ export class ContractDefinitionViewerComponent implements OnInit {
   searchText = '';
   private fetch$ = new BehaviorSubject(null);
 
-  constructor(private contractDefinitionService: ContractDefinitionService, private readonly dialog: MatDialog) {
+  constructor(private contractDefinitionService: ContractDefinitionService,
+              private notificationService: NotificationService,
+              private readonly dialog: MatDialog) {
   }
 
   ngOnInit(): void {
@@ -26,9 +32,9 @@ export class ContractDefinitionViewerComponent implements OnInit {
         switchMap(() => {
           const contractDefinitions$ = this.contractDefinitionService.getAllContractDefinitions();
           return !!this.searchText ?
-          contractDefinitions$.pipe(map(contractDefinitions => contractDefinitions.filter(contractDefinition => contractDefinition.id.toLowerCase().includes(this.searchText))))
+            contractDefinitions$.pipe(map(contractDefinitions => contractDefinitions.filter(contractDefinition => contractDefinition.id.toLowerCase().includes(this.searchText))))
             :
-          contractDefinitions$;
+            contractDefinitions$;
         }));
   }
 
@@ -37,7 +43,16 @@ export class ContractDefinitionViewerComponent implements OnInit {
   }
 
   onDelete(contractDefinition: ContractDefinitionDto) {
-    this.contractDefinitionService.deleteContractDefinition(contractDefinition.id).subscribe(() => this.fetch$.next(null));
+    const dialogData = ConfirmDialogModel.forDelete("contract definition", contractDefinition.id);
+
+    const ref = this.dialog.open(ConfirmationDialogComponent, {maxWidth: '20%', data: dialogData});
+
+    ref.afterClosed().subscribe(res => {
+      if (res) {
+        this.contractDefinitionService.deleteContractDefinition(contractDefinition.id).subscribe(() => this.fetch$.next(null));
+      }
+    });
+
   }
 
   onCreate() {
@@ -45,7 +60,9 @@ export class ContractDefinitionViewerComponent implements OnInit {
     dialogRef.afterClosed().pipe(first()).subscribe((result: { contractDefinition?: ContractDefinitionDto }) => {
       const newContractDefinition = result?.contractDefinition;
       if (newContractDefinition) {
-        this.contractDefinitionService.createContractDefinition(newContractDefinition).subscribe(() => this.fetch$.next(null));
+        this.contractDefinitionService.createContractDefinition(newContractDefinition).subscribe(() => this.fetch$.next(null),
+          error => this.notificationService.showError("Contract definition cannot be created"),
+          () => this.notificationService.showInfo("Contract definition created"));
       }
     });
   }
