+++ src/modules/edc-demo/components/policy-view/policy-view.component.ts
@@ -1,11 +1,11 @@
 import {Component, OnInit} from '@angular/core';
 import {Policy, PolicyService} from "../../../edc-dmgmt-client";
-import {BehaviorSubject, Observable, Observer, of, Subscriber} from "rxjs";
+import {BehaviorSubject, Observable, Observer, of} from "rxjs";
 import {first, map, switchMap} from "rxjs/operators";
 import {MatDialog} from "@angular/material/dialog";
 import {NewPolicyDialogComponent} from "../new-policy-dialog/new-policy-dialog.component";
 import {NotificationService} from "../../services/notification.service";
-import {HttpErrorResponse} from "@angular/common/http";
+import {ConfirmationDialogComponent, ConfirmDialogModel} from "../confirmation-dialog/confirmation-dialog.component";
 
 @Component({
   selector: 'app-policy-view',
@@ -30,6 +30,7 @@ export class PolicyViewComponent implements OnInit {
         this.notificationService.showInfo("Successfully completed")
       },
     }
+
   }
 
   ngOnInit(): void {
@@ -66,7 +67,16 @@ export class PolicyViewComponent implements OnInit {
   }
 
   delete(policy: Policy) {
-    this.policyService.deletePolicy(policy.uid).subscribe(this.errorOrUpdateSubscriber);
+
+    const dialogData = ConfirmDialogModel.forDelete("policy", policy.uid);
+
+    const ref = this.dialog.open(ConfirmationDialogComponent, {maxWidth: '20%', data: dialogData});
+
+    ref.afterClosed().subscribe(res => {
+      if (res) {
+        this.policyService.deletePolicy(policy.uid).subscribe(this.errorOrUpdateSubscriber);
+      }
+    });
   }
 
   private showError(error: Error) {
